Web_App_testing/requirements.txt
astroid==2.0.4
behave==1.2.6
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
comtypes==1.1.7
cryptography==2.4.2
Faker==1.0.0
idna==2.7
nerodia==0.14.0
parse==1.8.4
parse-type==0.4.2
temp-mail==0.2
PyYAML==3.13
requests==2.20.1
selenium==3.14.1
six==1.11.0
text-unidecode==1.2
urllib3==1.23
uuid==1.30
StringGenerator==0.3.1

